fix bugs in temporal locomo codes in evaluation
