+++ evaluation/scripts/temporal_locomo/locomo_eval.py
@@ -371,7 +371,7 @@ if __name__ == "__main__":
     parser.add_argument(
         "--version",
         type=str,
-        default="v0.2.1",
+        default="v1.0.1",
         help="Version identifier for loading results (e.g., 1010)",
     )
     parser.add_argument(
