+++ evaluation/scripts/temporal_locomo/temporal_locomo_eval.py
@@ -17,7 +17,7 @@ sys.path.insert(0, str(BASE_DIR))  # Enable execution from any working directory
 logger = get_logger(__name__)
 
 
-class TemporalLocomoEval(LocomoEvalModelModules):
+class TimeLocomoEval(LocomoEvalModelModules):
     def __init__(self, args):
         super().__init__(args=args)
 
@@ -25,7 +25,15 @@ class TemporalLocomoEval(LocomoEvalModelModules):
         self.locomo_processor = LocomoProcessor(args=args)
 
     def run_eval_pipeline(self):
-        # Check if temporal_locomo dataset exists, if not convert it
+        """
+        Run the complete evaluation pipeline including dataset conversion,
+        data ingestion, and processing.
+        """
+        print("=" * 80)
+        print("Starting TimeLocomo Evaluation Pipeline")
+        print("=" * 80)
+
+        # Step 1: Check if temporal_locomo dataset exists, if not convert it
         temporal_locomo_file = self.data_dir / "temporal_locomo" / "temporal_locomo_qa.json"
         if not temporal_locomo_file.exists():
             print(f"Temporal locomo dataset not found at {temporal_locomo_file}")
@@ -35,16 +43,38 @@ class TemporalLocomoEval(LocomoEvalModelModules):
         else:
             print(f"Temporal locomo dataset found at {temporal_locomo_file}, skipping conversion.")
 
-        # ingestion
+        # Step 2: Data ingestion
+        print("\n" + "=" * 50)
+        print("Step 2: Data Ingestion")
+        print("=" * 50)
         if not self.ingestion_storage_dir.exists() or not any(self.ingestion_storage_dir.iterdir()):
+            print(f"Directory {self.ingestion_storage_dir} not found, starting data ingestion...")
             self.locomo_ingestor.run_ingestion()
+            print("Data ingestion completed.")
         else:
             print(
                 f"Directory {self.ingestion_storage_dir} already exists and is not empty, skipping ingestion."
             )
 
-        # processing
+        # Step 3: Processing and evaluation
+        print("\n" + "=" * 50)
+        print("Step 3: Processing and Evaluation")
+        print("=" * 50)
+        print("Running locomo processing to search and answer...")
+
+        print("Starting locomo processing to generate search and response results...")
         self.locomo_processor.run_locomo_processing()
+        print("Processing completed successfully.")
+
+        # Step 4: Summary
+        print("\n" + "=" * 80)
+        print("Evaluation Pipeline Completed Successfully!")
+        print("=" * 80)
+        print("Results saved to:")
+        print(f"  - Search results: {self.search_path}")
+        print(f"  - Response results: {self.response_path}")
+        print(f"  - Statistics: {self.stats_path}")
+        print("=" * 80)
 
 
 if __name__ == "__main__":
@@ -59,7 +89,7 @@ if __name__ == "__main__":
     parser.add_argument(
         "--version",
         type=str,
-        default="v0.2.1",
+        default="v1.0.1",
         help="Version identifier for saving results (e.g., 1010)",
     )
     parser.add_argument(
@@ -70,5 +100,5 @@ if __name__ == "__main__":
     )
     args = parser.parse_args()
 
-    evaluator = TemporalLocomoEval(args=args)
+    evaluator = TimeLocomoEval(args=args)
     evaluator.run_eval_pipeline()
